Moving style calculator to separate file
